Examples, new instructions, new register, labels
Added support for labels in code. Labeled lines begin
with text and a colon and can be referenced using
%labelName:

```
main:    PUSH 0
         PUSH 1
         ADD
         POP
         JMP %main
```

A new register "RT" is added. It was originally for the
return location of a function call. Function calls aren't
yet final so I'm keeping it.

New Instructions:

- JMPLZ %label: Jump to label if TOS is less than 0
- JMPEQ %label: Jump to label if TOS equals 0
- JMPNEQ %label: Jump to label if TOS does not equal 0
- CALL %label: Jump to label and create a new call frame
- RET: Return to calling location
- PRINTS: Print the stack to stdout
- SWAP: Swap the top two stack items
- DUP: Duplicate TOS value
- MUL: Multiply the top two stack values return the result to TOS
- DIV: Divid the top two stack values return the result to TOS

Added a few examples of code using factorials and fibonacci series.