Added new opcodes, support strings, simpler stack
The stack is now an array with 1024 slots.

Registers and stack values may be a string or integer.
Opcodes are available for either.

A program is now a slice of bytes instead of a slice
of int64. It's easier to get strings this way.

Strings are up to 32768 bytes long.

New opcodes: PUSHSTR, SETSTR.

Renamed opcodes:

- PRINTS -> DUMP   (dump stack contents)
- PRINTR -> DUMPR  (print all registers)
-        -> PRINTR (print a given register)
- PUSH   -> PUSHI
- POPREG -> STORE
- SET    -> SETI

New flags:

- `-d`: Print debug output
- `-c`: Compile source file to bytecode, don't execute
- `-o`: Output for compiled file