Added POPREG code, fixed STORE command
The documentation said STORE didn't pop the stack
but the implementation did actually pop the stack
and the examples took advantage of that. A new
opcode has been added POPREG that will pop the stack
and store the value in a register. STORE now conforms
to the documentation and will store the TOS value
in a register, but will not pop the stack.