Modified SETI, CALL, and RETURN. Added PARAM and JUMPREG
SETI can now take a label which stores the label location in a register.

CALL sets the $RT register and uses a frame pointer.

RETURN uses the $RT register. The caller is responsible for restoring
the frame pointer.

PARAM $reg # will store the function parameter in position # to $reg.
Parameters are relative to the current frame pointer. Parameters need
to be pushed in reverse order.

JUMPREG will set $PC to the value in $reg.