Added CMP, JMPZ* instructions
CMP compares two registers and sets the zero flag
to -1, 0, or 1 depending on if the first register
is less than, equal to, or greater than, the second.

The new JMPZ* instructions work the same as the normal
jump instructions but instead use the value of the zero
flag to make a decision instead of the top of stack.